More comments and error handling

+++ higurashiInstaller.py
@@ -2,30 +2,53 @@ from __future__ import print_function, unicode_literals
 import sys, os, os.path as path, platform, shutil, glob, subprocess, json
 try:
 	from urllib.request import urlopen, Request
-except:
-	from urllib2 import urlopen, Request
+	from urllib.error import HTTPError
+except ImportError:
+	from urllib2 import urlopen, Request, HTTPError
 
+# Python 2 Compatibility
 try: input = raw_input
 except NameError: pass
 
+try:
+	"".decode("utf-8")
+	def decodeStr(string):
+		return string.decode("utf-8")
+except AttributeError:
+	def decodeStr(string):
+		return string
+
 def exitWithError():
+	# I think the windows command prompt immediately exits if you launched the program from a GUI
+	# So make sure they can read any error messages by waiting for input
 	if platform.system() == "Windows":
 		input()
 	sys.exit(1)
 
 class Installer:
 	def __init__(self, directory, info):
+		"""
+		Installer Init
+		
+		:param str directory: The directory of the game
+		:param dict info: The info dictionary from server JSON file for the requested target
+		"""
 		self.directory = directory
+		self.info = info
+
 		if platform.system() == "Darwin":
 			self.dataDirectory = path.join(self.directory, "Contents/Resources/Data")
 		else:
 			self.dataDirectory = path.join(self.directory, info["dataname"])
+
 		self.assetsDir = path.join(self.dataDirectory, "StreamingAssets")
-		self.info = info
+		
 		if path.exists(path.join(self.directory, "steam_api.dll")):
 			self.isSteam = True
 		else:
 			self.isSteam = False
+
+		self.downloadDir = info["name"] + "Download"
 
 		if path.exists("./7za"):
 			self.unzip = "./7za"
@@ -40,12 +63,19 @@ class Installer:
 			self.aria = "aria2c"
 
 	def backupUI(self):
+		"""
+		Backs up the `sharedassets0.assets` file
+		"""
 		uiPath = path.join(self.dataDirectory, "sharedassets0.assets")
 		backupPath = path.join(self.dataDirectory, "sharedassets0.assets.backup")
-		if not path.exists(backupPath):
+		if path.exists(uiPath) and not path.exists(backupPath):
 			os.rename(uiPath, backupPath)
 
 	def cleanOld(self):
+		"""
+		Removes folders that shouldn't persist through the install
+		(CompiledUpdateScripts, CG, and CGAlt)
+		"""
 		oldCG = path.join(self.assetsDir, "CG")
 		oldCGAlt = path.join(self.assetsDir, "CGAlt")
 		compiledScriptsPattern = path.join(self.assetsDir, "CompiledUpdateScripts/*.mg")
@@ -60,6 +90,9 @@ class Installer:
 			shutil.rmtree(oldCGAlt)
 
 	def download(self):
+		"""
+		Downloads the required files for the mod
+		"""
 		if platform.system() == "Windows":
 			try:
 				files = self.info["files"]["win"]
@@ -77,7 +110,7 @@ class Installer:
 				else:
 					files = self.info["files"]["unix-mg"]
 		try:
-			os.mkdir("Download")
+			os.mkdir(self.downloadDir)
 		except:
 			pass
 		fileList = open("downloadList.txt", "w")
@@ -102,12 +135,22 @@ class Installer:
 		os.remove("downloadList.txt")
 
 	def extractFiles(self):
-		for file in sorted(os.listdir("Download")):
-			subprocess.call([self.unzip, "x", path.join("Download", file), "-aoa"])
+		"""
+		Extracts downloaded files using 7zip
+		"""
+		for file in sorted(os.listdir(self.downloadDir)):
+			subprocess.call([self.unzip, "x", path.join(self.downloadDir, file), "-aoa"])
 
 	def moveFilesIntoPlace(self, fromDir=None, toDir=None):
+		"""
+		Moves files from the directory they were extracted to
+		to the game data folder
+
+		fromDir and toDir are for recursion, leave them at their defaults to start the process
+		"""
 		if fromDir is None: fromDir = self.info["dataname"]
 		if toDir is None: toDir = self.dataDirectory
+
 		for file in os.listdir(fromDir):
 			src = path.join(fromDir, file)
 			target = path.join(toDir, file)
@@ -122,44 +165,113 @@ class Installer:
 		os.rmdir(fromDir)
 
 	def cleanup(self):
-		shutil.rmtree("Download")
+		"""
+		General cleanup and other post-install things
 
-def getGameList():
-	file = urlopen(Request("http://07th-mod.com/installer/higurashi.json", headers={'User-Agent': 'Mozilla'}))
+		Removes downloaded 7z files
+		On mac, modifies the application Info.plist with new values if available
+		"""
+		try:
+			shutil.rmtree(self.downloadDir)
+		except OSError:
+			pass
+
+		if platform.system() == "Darwin":
+			# Allows fixing up application Info.plist file so that the titlebar doesn't show `Higurashi01` as the name of the application
+			# Can also add a custom CFBundleIdentifier to change the save directory (e.g. for Console Arcs)
+			infoPlist = path.join(self.directory, "Contents/Info.plist")
+			infoPlistJSON = subprocess.check_output(["plutil", "-convert", "json", "-o", "-", infoPlist])
+			parsed = json.loads(infoPlistJSON)
+			if "CFBundleName" in self.info and parsed["CFBundleName"] != self.info["CFBundleName"]:
+				subprocess.call(["plutil", "-replace", "CFBundleName", "-string", self.info["CFBundleName"], infoPlist])
+			if "CFBundleIdentifier" in self.info and parsed["CFBundleIdentifier"] != self.info["CFBundleIdentifier"]:
+				subprocess.call(["plutil", "-replace", "CFBundleIdentifier", "-string", self.info["CFBundleIdentifier"], infoPlist])
+
+def getModList():
+	"""
+	Gets the list of available mods from the 07th Mod server
+
+	:return: A list of mod info objects
+	:rtype: list[dict]
+	"""
+	try:
+		file = urlopen(Request("http://07th-mod.com/installer/higurashi.json", headers={"User-Agent": ""}))
+	except HTTPError as error:
+		print(error)
+		print("Couldn't reach 07th Mod Server to download patch info")
+		print("Note that we have blocked Japan from downloading (VPNs are compatible with this installer, however)")
+		exitWithError()
 	info = json.load(file)
 	file.close()
 	return info
 
 def findPossibleGames():
+	"""
+	If supported, searches the computer for things that might be Higurashi games
+	Currently only does things on Mac OS
+	TODO: Find ways to search for games on other platforms
+
+	:return: A list of things that might be Higurashi games
+	:rtype: list[str]
+	"""
 	if platform.system() == "Darwin":
 		return sorted(x for x in subprocess.check_output(["mdfind", "kind:Application", "Higurashi"]).decode("utf-8").split("\n") if x)
 	else:
 		return []
 
-def getGameInfo(game, gameList):
+def getGameInfo(game, modList):
+	"""
+	Gets the name of the given game, if there's a mod for it
+
+	:param str game: The path to the game
+	:param list[dict] modList: The list of available mods (used for finding game names)
+	:return: The name of the game, or None if no game was matched
+	:rtype: str or None
+	"""
 	if platform.system() == "Darwin":
-		info = subprocess.check_output(["plutil", "-convert", "json", "-o", "-", path.join(game, "Contents/Info.plist")])
-		parsed = json.loads(info)
-		name = parsed["CFBundleExecutable"] + "_Data"
+		try:
+			info = subprocess.check_output(["plutil", "-convert", "json", "-o", "-", path.join(game, "Contents/Info.plist")])
+			parsed = json.loads(info)
+			name = parsed["CFBundleExecutable"] + "_Data"
+		except (OSError, KeyError):
+			return None
 	else:
 		for file in os.listdir(game):
 			if file.startsWith("HigurashiEp"):
 				name = file
 
-	for game in gameList:
+	for game in modList:
 		if game["dataname"] == name:
 			return game["target"]
 	return None
 
-def findGames(gameList):
+def findGames(modList):
+	"""
+	Find moddable games
+	Uses findPossibleGames and therefore will support finding games on the same platforms it does
+	:param list[dict] modList: The list of available mods
+	:return: A list of games that were found and moddable
+	:rtype: list[str]
+	"""
 	possibleGames = findPossibleGames()
 	games = []
 	for game in possibleGames:
-		if getGameInfo(game, gameList) is not None:
+		if getGameInfo(game, modList) is not None:
 			games.append(game)
 	return games
 
 def promptChoice(choiceList, guiPrompt, textPrompt, canOther=True):
+	"""
+	Prompts the user to choose from a list
+	Currently supports using a choose GUI on mac OS and falls back to a CLI chooser on other OSes
+	TODO: Find ways to display choose GUIs on other OSes
+	:param list[str] choiceList: The list of choices
+	:param str guiPrompt: The prompt to use in GUI mode
+	:param str textPrompt: The prompt to use in CLI mode.  Note that the user will be directed to select from a list of integers representing options so please mention that.
+	:param bool canOther: Whether or not to give the user an Other option, which will instruct them to give a path to an application
+	:return: The string that the user selected, or if canOther is true, possibly a path that was not in the option list
+	:rtype: str
+	"""
 	choice = "Other"
 	if platform.system() == "Darwin":
 		withOther = choiceList + ["Other"] if canOther else choiceList
@@ -184,23 +296,27 @@ def promptChoice(choiceList, guiPrompt, textPrompt, canOther=True):
 			choice = inputted.strip()
 			if path.split(choice)[1].startswith("HigurashiEp"):
 				choice = path.split(choice)[0]
-	return choice.decode("utf-8")
+	return decodeStr(choice)
 
-def printSupportedGames(gameList):
+def printSupportedGames(modList):
+	"""
+	Prints a list of games that have mods available for them
+	:param list[dict] modList: The list of available mods
+	"""
 	print("Supported games:")
-	for game in set(x["target"] for x in gameList):
+	for game in set(x["target"] for x in modList):
 		print("  " + game)
 
 print("Getting latest mod info...")
-gameList = getGameList()
-foundGames = findGames(gameList)
+modList = getModList()
+foundGames = findGames(modList)
 gameToUse = promptChoice(foundGames, "Please choose a game to mod", "Please type the number of a game you would like to mod.")
-targetName = getGameInfo(gameToUse, gameList)
+targetName = getGameInfo(gameToUse, modList)
 if not targetName:
 	print(gameToUse + " does not appear to be a supported higurashi game.")
-	printSupportedGames(gameList)
+	printSupportedGames(modList)
 	exitWithError()
-possibleMods = [x for x in gameList if x["target"] == targetName]
+possibleMods = [x for x in modList if x["target"] == targetName]
 if len(possibleMods) > 1:
 	modName = promptChoice([x["name"] for x in possibleMods], "Please choose a mod to install", "Please type the number of the mod to install", canOther=False)
 	mod = [x for x in possibleMods if x["name"] == modName][0]
@@ -211,6 +327,7 @@ installer = Installer(gameToUse, mod)
 print("Downloading...")
 installer.download()
 print("Extracting...")
+installer.backupUI()
 installer.cleanOld()
 installer.extractFiles()
 print("Moving files into place...")
